8/07/2022
This will be my last changes before entering the Holidays where I will be taking a break.

Light Mode for Information page has been complete. File handling is complete.

I am now very close to finishing. These are my final things to complete:
Night mode for Information page (Mainly night mode for graph)
Warning message when user tries to enter breaks early that the Study time they did will not count as completed and not go into the file-handling.
Add alarm/a sound to notify the user the timer is finished when they do not have attention on the screen.