26/07/2022 - Part 2 (Graph)
Here I made the night mode version of the graph displaying my users study data. To achieve this, I created an IF function which changed the parameters of the graph image. I then changed when the graph image was created (when the information page was loaded) so that when the user selects night/light mode in the information page, it will refresh and the respective night/light mode graph image will show.