1/08/2022
Comments for MainLoop has been completed as well as structuring of Def functions.
A bug was also found in the information page where singular characters would take users data who had that character in it. A requirement was made to the name variable of at least 3 Characters to prevent this.